requirements.txt
Django==1.7.1
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.4
python-social-auth
dj-database-url
django-mptt
